separation into mrtextfield and mrtextarea and mrtextinput
