+++ src/core/entities/MRTextInputEntity.js
@@ -1,4 +1,4 @@
-import { getSelectionRects } from 'troika-three-text';
+import { getCaretAtPoint, getSelectionRects, Text } from 'troika-three-text';
 
 import * as THREE from 'three';
 
@@ -95,14 +95,6 @@ export class MRTextInputEntity extends MRTextEntity {
     }
 
     /**
-     * @function
-     * @description Updates the cursor position based on click and selection location.
-     */
-    updateCursorPosition() {
-        mrjsUtils.error.emptyParentFunction();
-    }
-
-    /**
      *
      * @param {event} event - the keydown event
      */
@@ -115,7 +107,17 @@ export class MRTextInputEntity extends MRTextEntity {
      * @param {event} event - the keydown event
      */
     handleMouseClick(event) {
-        mrjsUtils.error.emptyParentFunction();
+        // Convert isx position from world position to local:
+        // - make sure textObj has updated matrices so we're not calculating info wrong
+        // - note: textObj doesnt need sync
+        this.textObj.updateMatrixWorld(true);
+        const inverseMatrixWorld = new THREE.Matrix4().copy(this.textObj.matrixWorld).invert();
+        const localPosition = inverseMatrixWorld * event.worldPosition;
+
+        // update cursor position based on click
+        const caret = getCaretAtPoint(this.textObj.textRenderInfo, localPosition.x, localPosition.y);
+        this.hiddenInput.selectionStart = caret.charIndex;
+        this.updateCursorPosition();
     }
 
     /**
@@ -161,9 +163,6 @@ export class MRTextInputEntity extends MRTextEntity {
         // and calling them manually. This allows us to call
         // super.func() for event functions; otherwise, theyre
         // not accessible.
-        //
-        // TODO - is this a reasonable approach? is javascript
-        // really this clunky?
        
         // Blur events
         this.addEventListener('blur', () => {
@@ -195,7 +194,7 @@ export class MRTextInputEntity extends MRTextEntity {
             this.updateCursorPosition();
         });
         this.hiddenInput.addEventListener('keydown', (event) => {
-            console.log('THIS WAS A KEYDOWN');
+            console.log('handling EVENT, reg, KEYDOWN');
             this.handleKeydown(event);
         });
 
@@ -203,5 +202,96 @@ export class MRTextInputEntity extends MRTextEntity {
         this.addEventListener('update-cursor-position', () => {
             this.updateCursorPosition();
         });
+    }
+
+    /**
+     * @function
+     * @description Updates the cursor position based on click and selection location.
+     */
+    updateCursorPosition(fromCursorMove=false, isNewLine=false) {
+
+        const updateBasedOnSelectionRects = (cursorIndex, isNewLine) => {
+            // Setup variables for calculations.
+            // XXX - handle cursor position change for visible lines for scrolloffset here in future
+            const textBeforeCursor = this.hiddenInput.value.substring(0, cursorIndex);
+            const allLines = this.hiddenInput.value.split('\n');
+            const linesBeforeCursor = textBeforeCursor.split('\n');
+            const cursorIsOnLineIndex = linesBeforeCursor.length - 1;
+
+            let cursorXOffsetPosition = 0;
+            let cursorYOffsetPosition = 0;
+            
+            if (isNewLine) {
+                console.log('in the isNewLine section');
+                // Faking the newline cursor position since troika doesnt create it until you type something
+                // after the '\n' character sequence, which means the cursor lingers on the current line one
+                // keypress too long, which isnt expected.
+                //
+                // Note that for this case, the cursor indicating position is /not/ matching the selectionRects
+                // nor the this.hiddenInputs physical position.
+                cursorXOffsetPosition = 0;
+                cursorYOffsetPosition = 0 - this.cursorHeight * (cursorIsOnLineIndex + 1);
+                console.log('cursorXOffsetPosition: ', cursorXOffsetPosition, 'cursorYOffsetPosition: ', cursorYOffsetPosition);
+            } else {
+                // Assuming getSelectionRects is properly imported and used here
+                let selectionRects = getSelectionRects(this.textObj.textRenderInfo, textBeforeCursor.length - 1, cursorIndex);
+                if (selectionRects.length > 0) {
+                    let lastRect = selectionRects[selectionRects.length - 1];
+
+                    // Check if cursor matches our font size before using values.
+                    const cursorVisibleHeight = lastRect.top - lastRect.bottom;
+                    if (this.cursor.geometry.height != cursorVisibleHeight) {
+                        this.cursor.geometry.height = cursorVisibleHeight;
+                        this.cursor.geometry.needsUpdate = true;
+                        this.cursorHeight = cursorVisibleHeight;
+                    }
+
+                    // Add the cursor dimension info to the position s.t. it doesnt touch the text itself. We want
+                    // a little bit of buffer room.
+                    cursorXOffsetPosition = lastRect.right + this.cursorWidth;
+                    cursorYOffsetPosition = lastRect.bottom + this.cursorHeight;
+                }
+            }
+
+            // Update the cursor's 3D position
+            this.cursor.position.x = this.cursorStartingPosition.x + cursorXOffsetPosition;
+            this.cursor.position.y = this.cursorStartingPosition.y + cursorYOffsetPosition;
+            if (isNewLine) {
+                console.log('this.cursorStartingPosition:', this.cursorStartingPosition);
+                console.log('this.cursor.position:', this.cursor.position);
+            }
+            this.cursor.visible = true;
+        }
+
+        console.log('here in update cursor position');
+        // Check if we have any DOM element to work with.
+        if (!this.hiddenInput) {
+            return;
+        }
+
+        // Since no text is selected, this and selectionEnd are just the cursor position.
+        // XXX - when we actually allow for seleciton in future, some of the below will need to
+        // be thought through again.
+        const cursorIndex = this.hiddenInput.selectionStart;
+        console.log('found cursor index as:', cursorIndex);
+        
+        // early escape for empty text
+        if (cursorIndex == 0) {
+            this.cursor.position.x = this.cursorStartingPosition.x;
+            this.cursor.position.y = this.cursorStartingPosition.y;
+            this.cursor.visible = true;
+            return;
+        }
+
+        // Separating textObj sync from the cursor update based on rects
+        // since textObj sync resolves when there's actual changes to the
+        // object. Otherwise, it'll hang and never hit the update function.
+        if (fromCursorMove) {
+            updateBasedOnSelectionRects(cursorIndex, isNewLine);
+        } else {
+            this.textObj.sync(() => {
+                updateBasedOnSelectionRects(cursorIndex, isNewLine);
+            });
+        }
     }
 }
